backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "omnitrader-backend"
version = "0.1.0"
description = "OmniTrader Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ccxt==4.5.42",
    "aiohttp==3.13.3",
    "numpy==2.4.2",
    "pandas==3.0.1",
    "scipy==1.15.2",
    "TA-Lib==0.6.8",
    "pyyaml==6.0.3",
    "python-dotenv==1.2.2",
    "packaging==26.0",
    "structlog==25.5.0",
    "neo4j==6.1.0",
    "redis==6.4.0",
    "celery[redis]==5.6.2",
    "httpx==0.28.1",
    "psutil==7.0.0",
    "fastapi==0.135.1",
    "uvicorn[standard]==0.41.0",
    "feedparser>=6.0.0",
    "llama-index==0.12.21",
    "llama-index-llms-ollama==0.5.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<10.0.0",
    "pytest-asyncio>=0.21.0,<1.4.0",
    "ruff>=0.3.0,<1.0.0",
    "mypy>=1.0.0,<2.0.0",
    "pre-commit>=3.6.0,<4.0.0",
    "testcontainers>=3.0.0,<5.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [".git", ".venv", "__pycache__", "build", "dist"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:.*Copy-on-Write.*:pandas.errors.Pandas4Warning",
    "ignore::DeprecationWarning",
]
asyncio_default_fixture_loop_scope = "function"
